Add support for updating .terraform.lock.hcl
Terraform v0.14 will introduce a dependency lock file named
`.terraform.lock.hcl`.
https://discuss.hashicorp.com/t/terraform-0-14-the-dependency-lock-file/15696

The lock file is an implementation detail and it's hard to update it
with the tfupdate command. Because the hash value format seems to allow
future terraform versions to change the hash algorithm gradually. If
tfupdate updates the lock file, it will be difficult to support multiple
terraform versions without an unexpected the lock file drift.
We should use `terraform init -upgrade` to update the lock file.