Add an example for updating .terraform.lock.hcl
Terraform v0.14 will introduce a dependency lock file named
`.terraform.lock.hcl`.
https://discuss.hashicorp.com/t/terraform-0-14-the-dependency-lock-file/15696

The lock file is an implementation detail and it's hard to update it
with the tfupdate command. Because the hash value format seems to allow
future terraform versions to change the hash algorithm gradually. If
tfupdate updates the lock file, it will be difficult to support multiple
terraform versions without an unexpected the lock file drift.

Perhaps a recommended way is to use the terraform command to update the
lock file. However most CIs use linux and laptops may be mac or windows
and if you run terraform on multiple platforms, it requires to generate
hashed for all platforms to avoid a lock file drift. So I added an
example for it, but I found it's inefficient because it requires
duplicate downloads for providers on init. I tried to use a plugin cache
dir or a local filesystem mirror, but I couldn't figure out a way to
avoid duplicate downloads on init. If I use cache or mirror, it doesn't
record zip hashes, and it records them only when download providers from
origin. I think the easiest way to use Terraform v0.14 with tfupdate is
adding .terraform.lock.hcl to .gitignore.